# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AffectedDeviceResponse": {
        "description": "One device that received configuration from a change group.",
        "properties": {
          "device_mac": {
            "title": "Device Mac",
            "type": "string"
          },
          "device_name": {
            "default": "",
            "title": "Device Name",
            "type": "string"
          },
          "device_type": {
            "default": "",
            "title": "Device Type",
            "type": "string"
          },
          "site_mist_id": {
            "default": "",
            "title": "Site Mist Id",
            "type": "string"
          }
        },
        "required": [
          "device_mac"
        ],
        "title": "AffectedDeviceResponse",
        "type": "object"
      },
...
      "BaselineConfidence": {
        "description": "Confidence in the pre-change baseline used for impact comparison.",
        "enum": [
          "high",
          "medium",
          "low",
          "none"
        ],
        "title": "BaselineConfidence",
        "type": "string"
      },
...
      "ChangeEvidenceResponse": {
        "description": "One deterministic statement supporting a group's assessment.",
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/ImpactSeverity",
            "default": "none"
          }
        },
        "required": [
          "label"
        ],
        "title": "ChangeEvidenceResponse",
        "type": "object"
      },
      "ChangeGroupDetailResponse": {
        "description": "One change group with its evidence, objects, and competing changes.",
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "affected_devices": {
            "items": {
              "$ref": "#/components/schemas/AffectedDeviceResponse"
            },
            "title": "Affected Devices",
            "type": "array"
          },
          "affected_site_ids": {
            "items": {
              "type": "string"
            },
            "title": "Affected Site Ids",
            "type": "array"
          },
          "audit_id": {
            "title": "Audit Id",
            "type": "string"
          },
          "baseline_confidence": {
            "$ref": "#/components/schemas/BaselineConfidence",
            "default": "none"
          },
          "changed_objects": {
            "items": {
              "$ref": "#/components/schemas/ChangedObjectResponse"
            },
            "title": "Changed Objects",
            "type": "array"
          },
          "competing_change_group_ids": {
            "items": {
              "type": "string"
            },
            "title": "Competing Change Group Ids",
            "type": "array"
          },
          "degraded_metrics": {
            "items": {
              "type": "string"
            },
            "title": "Degraded Metrics",
            "type": "array"
          },
          "deterministic_assessment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deterministic Assessment"
          },
          "device_count": {
            "title": "Device Count",
            "type": "integer"
          },
          "devices_label": {
            "title": "Devices Label",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ChangeEvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "impact_label": {
            "title": "Impact Label",
            "type": "string"
          },
          "impact_severity": {
            "$ref": "#/components/schemas/ImpactSeverity",
            "default": "none"
          },
          "is_mine": {
            "default": false,
            "title": "Is Mine",
            "type": "boolean"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/ChangeMetricResponse"
            },
            "title": "Metrics",
            "type": "array"
          },
          "monitoring_session_ids": {
            "items": {
              "type": "string"
            },
            "title": "Monitoring Session Ids",
            "type": "array"
          },
          "object_count": {
            "title": "Object Count",
            "type": "integer"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "recovery_state": {
            "$ref": "#/components/schemas/RecoveryState",
            "default": "not_applicable"
          },
          "source": {
            "$ref": "#/components/schemas/ChangeSource",
            "default": "webhook"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "audit_id",
          "occurred_at",
          "title",
          "summary",
          "object_count",
          "device_count",
          "devices_label",
          "impact_label"
        ],
        "title": "ChangeGroupDetailResponse",
        "type": "object"
      },
      "ChangeGroupListResponse": {
        "description": "One page of change groups and the total matching the query.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChangeGroupSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ChangeGroupListResponse",
        "type": "object"
      },
      "ChangeGroupSummaryResponse": {
        "description": "One change group as the Overview feed and Changes table render it.",
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "affected_site_ids": {
            "items": {
              "type": "string"
            },
            "title": "Affected Site Ids",
            "type": "array"
          },
          "audit_id": {
            "title": "Audit Id",
            "type": "string"
          },
          "degraded_metrics": {
            "items": {
              "type": "string"
            },
            "title": "Degraded Metrics",
            "type": "array"
          },
          "device_count": {
            "title": "Device Count",
            "type": "integer"
          },
          "devices_label": {
            "title": "Devices Label",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "impact_label": {
            "title": "Impact Label",
            "type": "string"
          },
          "impact_severity": {
            "$ref": "#/components/schemas/ImpactSeverity",
            "default": "none"
          },
          "is_mine": {
            "default": false,
            "title": "Is Mine",
            "type": "boolean"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/ChangeMetricResponse"
            },
            "title": "Metrics",
            "type": "array"
          },
          "monitoring_session_ids": {
            "items": {
              "type": "string"
            },
            "title": "Monitoring Session Ids",
            "type": "array"
          },
          "object_count": {
            "title": "Object Count",
            "type": "integer"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "recovery_state": {
            "$ref": "#/components/schemas/RecoveryState",
            "default": "not_applicable"
          },
          "source": {
            "$ref": "#/components/schemas/ChangeSource",
            "default": "webhook"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "audit_id",
          "occurred_at",
          "title",
          "summary",
          "object_count",
          "device_count",
          "devices_label",
          "impact_label"
        ],
        "title": "ChangeGroupSummaryResponse",
        "type": "object"
      },
      "ChangeMetricResponse": {
        "description": "One metric tile rendered on an Overview feed card.",
        "properties": {
          "from": {
            "default": "",
            "title": "From",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/ImpactSeverity",
            "default": "none"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "label",
          "value"
        ],
        "title": "ChangeMetricResponse",
        "type": "object"
      },
      "ChangeSource": {
        "description": "How the application learned about a change.",
        "enum": [
          "webhook",
          "reconcile",
          "restore",
          "snapshot"
        ],
        "title": "ChangeSource",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImpactSeverity": {
        "description": "Deterministic or AI-assisted impact rating.",
        "enum": [
          "none",
          "info",
          "warning",
          "critical"
        ],
        "title": "ImpactSeverity",
        "type": "string"
      },
...
      "MonitoringSessionListResponse": {
        "description": "Paginated monitoring session collection.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MonitoringSessionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "MonitoringSessionListResponse",
        "type": "object"
      },
      "MonitoringSessionResponse": {
        "description": "Complete monitoring session with no credentials or raw payloads.",
        "properties": {
          "ai_assessment": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Assessment"
          },
          "ai_assessment_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Assessment Error"
          },
          "audit_ids": {
            "items": {
              "type": "string"
            },
            "title": "Audit Ids",
            "type": "array"
          },
          "baseline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SleObservationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "config_applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Applied At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "degraded_metrics": {
            "items": {
              "type": "string"
            },
            "title": "Degraded Metrics",
            "type": "array"
          },
          "deterministic_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deterministic Summary"
          },
          "device_mac": {
            "title": "Device Mac",
            "type": "string"
          },
          "device_name": {
            "title": "Device Name",
            "type": "string"
          },
          "device_type": {
            "title": "Device Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "impact_severity": {
            "$ref": "#/components/schemas/ImpactSeverity"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/MonitoringIncidentResponse"
            },
            "title": "Incidents",
            "type": "array"
          },
          "monitoring_ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monitoring Ends At"
          },
          "monitoring_started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monitoring Started At"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/SleObservationResponse"
            },
            "title": "Observations",
            "type": "array"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MonitoringStatus"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "audit_ids",
          "site_id",
          "device_mac",
          "device_name",
          "device_type",
          "status",
          "baseline",
          "observations",
          "incidents",
          "config_applied_at",
          "monitoring_started_at",
          "monitoring_ends_at",
          "impact_severity",
          "deterministic_summary",
          "degraded_metrics",
          "ai_assessment",
          "ai_assessment_error",
          "warnings",
          "created_at",
          "completed_at"
        ],
        "title": "MonitoringSessionResponse",
        "type": "object"
      },
      "MonitoringStatus": {
        "description": "Monitoring lifecycle.",
        "enum": [
          "awaiting_config",
          "monitoring",
          "completed",
          "failed"
        ],
        "title": "MonitoringStatus",
        "type": "string"
      },
...
      "PydanticObjectId": {
        "example": "5eb7cf5a86d9755df3a6c593",
        "maxLength": 24,
        "minLength": 24,
        "pattern": "^[0-9a-f]{24}$",
        "type": "string"
      },
...
      "RecoveryState": {
        "description": "Whether an impacting change has recovered.",
        "enum": [
          "not_applicable",
          "monitoring",
          "recovered",
          "unrecovered",
          "completed"
        ],
        "title": "RecoveryState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CsrfToken": {
        "description": "Double-submit CSRF token. Required on every unsafe method of a cookie session.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      },
      "SessionCookie": {
        "description": "Browser session issued by `POST /auth/login`. Unsafe methods must also send the `cg_csrf` cookie value in the `X-CSRF-Token` header.",
        "in": "cookie",
        "name": "cg_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "summary": "Configuration history, point-in-time recovery, and impact monitoring for Juniper Mist.",
    "title": "Mist Config Guardian",
    "version": "0.2.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organizations/{organization_id}/change-groups": {
      "get": {
        "description": "List change groups in the selected window, newest first.",
        "operationId": "list_change_groups_api_v1_organizations__organization_id__change_groups_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "default": "24h",
              "enum": [
                "24h",
                "7d",
                "30d"
              ],
              "title": "Range",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "default": "any",
              "enum": [
                "any",
                "critical",
                "warning",
                "none"
              ],
              "title": "Severity",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeGroupListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "List Change Groups",
        "tags": [
          "Change Groups"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/change-groups/{change_group_id}": {
      "get": {
        "description": "Return one change group with its objects, devices, and evidence.",
        "operationId": "read_change_group_api_v1_organizations__organization_id__change_groups__change_group_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "change_group_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeGroupDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Read Change Group",
        "tags": [
          "Change Groups"
        ]
      }
    },
...
    "/api/v1/organizations/{organization_id}/monitoring": {
      "get": {
        "description": "List newest monitoring windows for one organization.",
        "operationId": "list_monitoring_sessions_api_v1_organizations__organization_id__monitoring_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MonitoringStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ImpactSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringSessionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "List Monitoring Sessions",
        "tags": [
          "Impact Monitoring"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/monitoring/{session_id}": {
      "get": {
        "description": "Return one organization-scoped monitoring session.",
        "operationId": "get_monitoring_session_api_v1_organizations__organization_id__monitoring__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Get Monitoring Session",
        "tags": [
          "Impact Monitoring"
        ]
      }
    },
...
  },
...
}